requirements.txt
tensorflow-macos
tensorflow-metal
numpy
pandas
matplotlib
scikit-learn
opencv-python-headless
streamlit
kaggle
